requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
easydict==1.13
executing==2.1.0
filelock==3.16.1
fonttools==4.54.0
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.25.1
idna==3.10
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
lightning-utilities==0.11.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.3
timm==1.0.9
torch==2.0.1
torchmetrics==1.4.2
torchvision==0.15.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
umap==0.1.1
urllib3==2.2.3
wandb==0.18.1
wcwidth==0.2.13
yarl==1.12.1
